requirements.txt
docx2pdf==0.1.8
docxcompose==1.3.4
docxtpl==0.16.0
num2words==0.5.10
numpy==1.23.0
openpyxl==3.0.10
pandas==1.4.3
pdfkit==1.0.0
streamlit==1.10.0

